[SWS-105] 장바구니 아이템 수정 및 삭제

+++ app/src/main/java/com/ite/sws/domain/cart/data/GetCartItemRes.kt
@@ -13,7 +13,7 @@ package com.ite.sws.domain.cart.data
  * </pre>
  */
 data class GetCartItemRes (
-    val cartId: Int,
+    val cartId: Long,
     val totalPrice: Int,
     val items: List<CartItem>
 )
